[store] Rename builder to create-store.
